* Fixed methods which uses 'int' type parameters for index to use 'size_t'.

+++ include/usql/SQLNode.h
@@ -155,7 +155,7 @@ public:
     return &children;
   }
 
-  SQLNode *getChildNode(int index);  
+  SQLNode *getChildNode(size_t index);  
   SQLNode *findChildNodeByType(int type);
 
   void clearChildNodes()
